refactor: move Span to parser module
